pyproject.toml
[tool.poetry]
name = "vicis-groove"
version = "0.1.0"
description = "A friends said that want to see how something like this works,so I decided to try make it. It's just get some musics and creates a playlist, kinda simple."
authors = ["Vici <telegram: @Vici83>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
